Moved grouping of items back to publishview
